requirements.txt
# LLM
groq>=0.4.0

# Telegram Bot
python-telegram-bot>=20.0

# LinkedIn API
requests>=2.31.0
requests-oauthlib>=1.3.1

# Image Generation
huggingface-hub>=0.19.0

# Research
googlesearch-python>=1.2.3
feedparser>=6.0.10
beautifulsoup4>=4.12.0

# Scheduler
APScheduler>=3.10.0

# CLI
rich>=13.0.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0
pyyaml>=6.0
